app/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-compute==1.15.0
slack-sdk==3.27.1


